Feature/bands service

+++ Isobar.Fm.Api/Isobar.Fm.Api/Models/Track.cs
@@ -10,6 +10,6 @@ namespace Isobar.Fm.Api.Models
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public int Duration { get; set; }
+        public Int64 Duration { get; set; }
     }
 }
